REL: add Trove classifiers
indicate that this version is compatible only with Python 2.
The reason is the presence of `\u` within strings within the
function `nx2tikz.nx2tikz._preamble`.

The escape sequence `\u` is not active in the default Python 2
strings (because they are not unicode strings), but is active
in the default Python 3 strings (because they are unicode strings).

So the string literal `'\usepackage{{tikz}}'` raises an error
in Python 3, due to the `\u` sequence not being followed by
the code of a unicode character. The error is:

```
SyntaxError: (unicode error) 'unicodeescape' codec can't
decode bytes in position 0-1: truncated \uXXXX escape
```

in contrast, the string literal b`'\usepackage{{tikz}}'` does
not raise an error in Python 3. In comparison, the string
literal `'\usepackage{{tikz}}'` does not raise an error in
Python 2, but the string literal `u'\usepackage{{tikz}}'` does.

[1] https://docs.python.org/3/reference/lexical_analysis.html
[2] https://docs.python.org/3/library/codecs.html#encodings-and-unicode
[3] https://docs.python.org/2/reference/lexical_analysis.html
[4] https://docs.python.org/2/library/codecs.html#encodings-and-unicode