MAI: avoid unintentionally forming unicode escape sequence
The sequence `\u` does not raise an error in default string
literals in Python 2, because those are not unicode strings.
But the sequence `\u` does raise an error in default string
literals in Python 3, if followed by characters that do not
form a unicode character code.

Escaping the backslash avoids this error in Python 3.

Read also the message of commit:
    e114a90f12a63e61a4daf3fae3ea5111f4975302

- <https://docs.python.org/3/reference/lexical_analysis.html>
- <https://docs.python.org/3/library/codecs.html#encodings-and-unicode>
- <https://docs.python.org/2/reference/lexical_analysis.html>
- <https://docs.python.org/2/library/codecs.html#encodings-and-unicode>