REL: add Trove classifiers for Python 3.7 to 3.9
following the update to the API of `networkx >= 2.0.0`,
which enables using `nx2tikz` with `networkx >= 2.4`,
which is required to use `nx2tikz` on Python 3.9.

Including the Trove classifiers for Python 3.7, 3.8
is based on the Trove classifiers listed for
[`networkx == 2.4`](
    https://pypi.org/project/networkx/2.4/)

Including the Trove classifier for Python 3.9 is based
on the observation that `nx2tikz` works with
`networkx >= 2.4` without errors on Python 3.9.

Read also the commit with title:
> REL: rm upper bound `networkx < 2.0.0` from `install_requires`